+++ ecowitt2mqtt/helpers/calculator/__init__.py
