+++ ecowitt2mqtt/helpers/calculator/battery.py
