+++ ecowitt2mqtt/helpers/calculator/meteo.py
