--- tests/util/__init__.py
@@ -1,1 +1,0 @@
-"""Define tests for utilities."""
